requirements.txt
pyside6==6.10.1
pyinstaller==6.17.0
bcrypt==5.0.0
python-dotenv==1.2.1
rich==14.2.0
sqlalchemy==2.0.44
aiosqlite==0.21.0
Pillow==10.4.0
pytesseract==0.3.13
PyPDF2==3.0.1
